requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.5
djangorestframework==3.16.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
